pyproject.toml
[tool.poetry]
name = "docbot"
version = "0.1.0"
description = "Bot using an LLM for retrieving and synthesizing documentation"
authors = ["Mat Leonard <leonard.mat@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
langchain = "^0.0.166"
aim = "^3.17.4"
python-dotenv = "^1.0.0"
openai = "^0.27.6"
chromadb = "^0.3.22"
google-search-results = "^2.4.2"
aioboto3 = "^11.2.0"
deeplake = "^3.4.3"
requests = "^2.30.0"
beautifulsoup4 = "^4.12.2"
markdownify = "^0.11.6"
unstructured = "^0.6.6"
tiktoken = "^0.4.0"
streamlit = "^1.22.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
mypy = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
